Create Log.tmLanguage
Also, if you wanted to include it, I made a basic syntax highlighting file specifically for the .log files.  It's basically the same as the unrealscript highlighting with the added keywords: warning, summary, success.  The only annoying thing about it is the escaped character highlighting.  It might be a setting within Sublime itself.  Not sure how to get rid of it.